+++ source/ci_source/providers/index.ts
@@ -1,3 +1,4 @@
+import { AppCenter } from "./AppCenter"
 import { Bitrise } from "./Bitrise"
 import { BuddyBuild } from "./BuddyBuild"
 import { Buildkite } from "./Buildkite"
@@ -42,6 +43,7 @@ const providers = [
   Netlify,
   CodeBuild,
   Codefresh,
+  AppCenter,
 ]
 
 // Mainly used for Dangerfile linting
@@ -65,6 +67,7 @@ const realProviders = [
   Netlify,
   CodeBuild,
   Codefresh,
+  AppCenter,
 ]
 
 export { providers, realProviders }