+++ CHANGELOG.md
@@ -14,7 +14,9 @@
 ## Master
 
 <!-- Your comment below this -->
+
 - Add support for AppCenter - [@mrndjo]
+- Fix structuredDiffForFile for BitBucket Server - [@osmestad]
 
 # 7.1.4
 
@@ -1618,6 +1620,7 @@ Not usable for others, only stubs of classes etc. - [@orta]
 [@nminhnguyen]: https://github.com/NMinhNguyen
 [@nornagon]: https://github.com/nornagon
 [@orta]: https://github.com/orta
+[@osmestad]: https://github.com/osmestad
 [@patrickkempff]: https://github.com/patrickkempff
 [@peterjgrainger]: https://github.com/peterjgrainger
 [@randak]: https://github.com/randak