+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 <!-- Your comment below this -->
 - Add support for AppCenter - [@mrndjo]
 
+- Adds GitLab & GitLab CI support. - [@notjosh], [@bigkraig], [@jamime]
+
 # 7.1.4
 
 - Un-hardcodes the repo in `danger.github.utils.createOrUpdatePR`- [@ds300]
@@ -1634,3 +1636,4 @@ Not usable for others, only stubs of classes etc. - [@orta]
 [@dblandin]: https://github.com/dblandin
 [@paulmelnikow]: https://github.com/paulmelnikow
 [@ds300]: https://github.com/ds300
+[@jamime]: https://github.com/jamime